Simplify writing of UUIDs
